requirements.txt
Flask==2.0.2
Flask-Cors==3.0.10
python-dotenv==0.19.2
simple-websocket==0.5.0
python-socketio==5.5.1
python-dotenv==0.19.2
flask-marshmallow==0.14.0
flask_sqlalchemy==2.5.1
pymysql==1.0.2
cryptography==36.0.1
marshmallow-sqlalchemy==0.27.0
pyjwt==2.3.0
requests==2.27.1
python-dateutil==2.8.2